API security and authorization implementation

+++ src/routers/task_router.js
@@ -1,12 +1,18 @@
 const express = require('express')
 const task_router = express.Router()
+const auth = require("../middleware/auth")
 const task_model = require('../models/task')
 
 // ------------------- TASK
 
 // create a task
-task_router.post('/tasks', async (req, res) => {
-  const task = new task_model(req.body)
+task_router.post('/tasks', auth, async (req, res) => {
+  const task = new task_model(
+    {
+      ...req.body,
+      owner: req.user._id
+    }
+  )
 
   try {
     await task.save()
@@ -17,21 +23,25 @@ task_router.post('/tasks', async (req, res) => {
 
 })
 // get all tasks
-task_router.get('/tasks', async (req, res) => {
+task_router.get('/tasks', auth, async (req, res) => {
   try {
-    const tasks = await task_model.find({})
-    res.status(200).send(tasks)
+    // const tasks = await task_model.find({ owner: req.user._id})
+
+    await req.user.populate("usertasks")
+    res.status(200).send(req.user.usertasks)
   } catch (error) {
     res.status(400).send()
   }
 })
 
 // get a single task
-task_router.get('/tasks/:taskId', async (req, res) => {
-  const task_id = req.params.taskId
+task_router.get('/tasks/:taskId', auth, async (req, res) => {
+
+  const _id = req.params.taskId
 
   try {
-    const task = await task_model.findById(task_id)
+    const task = await task_model.findOne({ _id, owner: req.user._id })
+
     if (!task) { res.status(404).send() }
     res.status(200).send(task)
   } catch (error) {
@@ -39,8 +49,8 @@ task_router.get('/tasks/:taskId', async (req, res) => {
   }
 })
 
-// update a tas
-task_router.patch("/tasks/:taskid", async (req, res) => {
+// update a task
+task_router.patch("/tasks/:taskid", auth, async (req, res) => {
 
   const updates = Object.keys(req.body)
   const allowedUpdates = ["description", "completed"]
@@ -51,10 +61,14 @@ task_router.patch("/tasks/:taskid", async (req, res) => {
   }
 
   try {
-    const task = await task_model.findByIdAndUpdate(req.params.taskid, req.body, { new: true, runValidators: true })
+    const task = await task_model.findOne({ _id: req.params.taskid, owner: req.user._id })
+
     if (!task) {
       return res.status(404).send({ error: "Task not found! Please create one." })
     }
+
+    updates.forEach((update) => task[update] = req.body[update])
+    await task.save()
     res.status(200).send(task)
   } catch (error) {
     res.status(500).send(error)
@@ -63,16 +77,16 @@ task_router.patch("/tasks/:taskid", async (req, res) => {
 })
 
 // delete task
-task_router.delete("/tasks/:id", async (req, res) => {
+task_router.delete("/tasks/:id", auth, async (req, res) => {
 
   try {
-    const task = await task_model.findByIdAndDelete(req.params.id)
+    const task = await task_model.findByIdAndDelete({ _id: req.params.id, owner: req.user._id })
     if (!task) {
       return res.status(404).send({ error: "No task found!" })
     }
     res.status(200).send(task)
   } catch (error) {
-    res.status(500).send({ error})
+    res.status(500).send({ error })
   }
 
 })
